+++ OpenDreamRuntime/DreamServer.cs
@@ -9,19 +9,19 @@ namespace OpenDreamVM {
     public delegate void DreamConnectionReadyEventHandler(DreamConnection connection);
     public abstract class DreamServer {
         // Interface
-        public abstract void Start();
+        public abstract void Start(DreamRuntime runtime);
         public abstract void Process();
 
         // Implementation
         public readonly List<DreamConnection> Connections = new();
-        public event DreamConnectionReadyEventHandler DreamConnectionRequest;
+        public abstract event DreamConnectionReadyEventHandler DreamConnectionRequest;
 
         protected Dictionary<PacketID, Action<DreamConnection, IPacket>> _packetIDToCallback = new Dictionary<PacketID, Action<DreamConnection, IPacket>>();
 
-        public readonly IPAddress Address;
-        public readonly int Port;
+        public IPAddress Address { protected set; get; }
+        public int Port { protected set; get; }
 
-        internal void RegisterPacketCallback<PacketClass>(PacketID packetID, Action<DreamConnection, PacketClass> packetCallback) where PacketClass : IPacket, new() {
+        public void RegisterPacketCallback<PacketClass>(PacketID packetID, Action<DreamConnection, PacketClass> packetCallback) where PacketClass : IPacket, new() {
             if (_packetIDToCallback.ContainsKey(packetID)) throw new Exception("Packet ID '" + packetID + "' already has a callback");
 
             if (packetCallback != null) {
