--- OpenDreamServer/Dream/DreamGlobalVariable.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace OpenDreamServer.Dream {
-    class DreamGlobalVariable {
-        public DreamValue Value;
-
-        public DreamGlobalVariable(DreamValue value) {
-            Value = value;
-        }
-    }
-}
