+++ OpenDreamServer/OpenDreamServer.csproj
@@ -5,6 +5,8 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
-  <Import Project="..\OpenDreamShared\OpenDreamShared.projitems" Label="Shared" />
-
+  <ItemGroup>
+    <ProjectReference Include="..\OpenDreamRuntime\OpenDreamRuntime.csproj" />
+  </ItemGroup>
+  
 </Project>
