--- OpenDreamServer/Resources/ConsoleOutputResource.cs
@@ -1,22 +1,0 @@
-﻿using OpenDreamServer.Dream;
-using System;
-
-namespace OpenDreamServer.Resources {
-    //A special resource that outputs to the console
-    //world.log defaults to this
-    class ConsoleOutputResource : DreamResource {
-        public ConsoleOutputResource() : base(null, null) { }
-
-        public override string ReadAsString() {
-            return null;
-        }
-
-        public override void Output(DreamValue value) {
-            if (value.TryGetValueAsString(out string text)) {
-                Console.WriteLine(text);
-            } else {
-                Console.WriteLine(value.ToString());
-            }
-        }
-    }
-}
